AMI locked to Amazon Linux 2

+++ ec2mc/commands/user_sub/rotate_key_cmd.py
@@ -1,6 +1,5 @@
 from ec2mc import consts
 from ec2mc.utils import aws
-from ec2mc.utils import halt
 from ec2mc.utils import os2
 from ec2mc.utils.base_classes import CommandBase
 from ec2mc.validate import validate_perms
@@ -11,49 +10,21 @@ class RotateUserKey(CommandBase):
         """delete IAM user's access key(s) and create new one"""
         self.iam_client = aws.iam_client()
         path_prefix = f"/{consts.NAMESPACE}/"
-        user_name = cmd_args['name']
+        user_name = aws.validate_user_exists(path_prefix, cmd_args['name'])
 
-        iam_users = self.iam_client.list_users(PathPrefix=path_prefix)['Users']
-        for iam_user in iam_users:
-            if iam_user['UserName'].lower() == user_name.lower():
-                user_name = iam_user['UserName']
-                break
-        else:
-            halt.err(f"IAM user \"{user_name}\" not found from AWS.")
+        old_access_keys = self.iam_client.list_access_keys(
+            UserName=user_name)['AccessKeyMetadata']
+        old_key_ids = [key['AccessKeyId'] for key in old_access_keys]
 
-        if user_name == consts.IAM_NAME:
-            new_key, old_key_ids = self.rotate_own_access_key(user_name)
-        else:
-            new_key, old_key_ids = self.rotate_user_access_key(user_name)
+        new_key = self.rotate_user_key(old_key_ids, user_name)
         self.update_config_dict(new_key, old_key_ids)
 
         print("")
         print(f"{user_name}'s access key rotated.")
 
 
-    def rotate_user_access_key(self, user_name):
-        """rotate access key for IAM user other than default IAM user"""
-        old_access_keys = self.iam_client.list_access_keys(
-            UserName=user_name)['AccessKeyMetadata']
-        old_key_ids = [key['AccessKeyId'] for key in old_access_keys]
-        for key_id in old_key_ids:
-            self.iam_client.delete_access_key(
-                UserName=user_name,
-                AccessKeyId=key_id
-            )
-
-        new_key = self.iam_client.create_access_key(
-            UserName=user_name)['AccessKey']
-        new_key = {new_key['AccessKeyId']: new_key['SecretAccessKey']}
-
-        return new_key, old_key_ids
-
-
-    def rotate_own_access_key(self, user_name):
-        """rotate access key for default IAM user"""
-        old_access_keys = self.iam_client.list_access_keys(
-            UserName=user_name)['AccessKeyMetadata']
-        old_key_ids = [key['AccessKeyId'] for key in old_access_keys]
+    def rotate_user_key(self, old_key_ids, user_name):
+        """rotate IAM user's access key"""
         for key_id in old_key_ids:
             if key_id != consts.KEY_ID:
                 self.iam_client.delete_access_key(
@@ -64,13 +35,14 @@ class RotateUserKey(CommandBase):
         new_key = self.iam_client.create_access_key(
             UserName=user_name)['AccessKey']
         new_key = {new_key['AccessKeyId']: new_key['SecretAccessKey']}
-        self.iam_client.delete_access_key(
-            UserName=user_name,
-            AccessKeyId=consts.KEY_ID
-        )
 
-        return new_key, old_key_ids
+        if consts.KEY_ID in old_key_ids:
+            self.iam_client.delete_access_key(
+                UserName=user_name,
+                AccessKeyId=consts.KEY_ID
+            )
 
+        return new_key
 
     @staticmethod
     def update_config_dict(new_key, old_key_ids):
